testapp/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.4
django-redis==5.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
packaging==23.1
psycopg2-binary==2.9.7
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
rpds-py==0.9.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
